Sends an InputEvent over the network
This is the first commit in the networktest branch. The commits in this
branch are only to test networking capability. Most commits are quick
prototypes and are, therefore, very messy.

The EventProcessor will request for the address to send the InputEvent
to by spawning a window. The InputEvent is then packed into a datagram
and sent across the network.

Made all InputStrings in InputEvent two characters long to make it
easier to pack into the datagram.