Rearranges package hierarchy
The network infrastructure will use a client server model. The overall
design will be a MVC. The client will send an input event to the model
which will alter the games state. The model will then send several
sprites to the viewer which will then render it on the screen. This
commit reorders the files to reflect this setup.

GameEngine has been renamed to GameLoop to better represent it's actual
purpose.