Controller talks to model over loopback
Instead of the controller directly adding InputEvents to the InputQueue
it turns the InputEvent into a DatagramPacket which is then sent over
the loopback on port 1729 to the model. The model converts the
DatagramPacket back into an InputEvent which is then added to the
InputQueue.

Eventually the model (server) will be able to run on a different
computer and the controller will send it network packets.