Serialize entity instead of writing to ByteBuffer
The UDP packets sent to the ViewReceiver from the ViewSender consisted
of an entity object that was serialized by manually writing data to a
ByteBuffer (mainly because I didn't know of java.io.Serializable).

When an entity write to the ByteBuffer it cleared the drawEventQueue as
a side effect (by popping the elements onto the ByteBuffer). Since the
program does not write to a ByteBuffer Entity.clearDrawEvents() is
called after every packet is sent.

Several methods were updated to take an Entity as an argument instead of
a ByteBuffer.