Serialize InputEvent; do not write to ByteBuffer
The UDP packets sent to InputEventReceiver from the InputEventSender
consited of an InputEvent that was serialized by manually writing data
to a ByteBuffer.

This made the handshake very awkward because InputEventReciever would
have to check if the UDP packet was a String or a InputEvent.

A serializable HandShake class is now sent to the InputEventReceiver as
well as a serialized InputEvent.

This changed removed all calls to the UTF8 class. It was thus removed.